requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.0
huggingface_hub>=0.23.0
